+++ lua/revelcommon/enums/KnifeSubtype.lua
@@ -1,3 +1,5 @@
 return {
+    DEFAULT = 0,
+    PROJECTILE = 1, -- Example: brimstone + mom's knife
     SWING = 4,
 }